+++ R/web_table.R
@@ -8,8 +8,10 @@
 #' @param rnames Row names.
 #' @param buttons Buttons: "excel", "copy" or "none". Default c("excel", "copy")
 #' @param file_name Excel file name
-#' @param scrolly Windows height to show the table. Default "60vh"
-#'
+#' @param scrolly Windows height to show the table. Default "45vh"
+#' @param columnwidth Column width. Default '200px'
+#' @param width Width in pixels or percentage (Defaults to automatic sizing)
+#' 
 #' @return table in markdown format for html documents
 #'
 #' @importFrom dplyr mutate across
@@ -35,6 +37,8 @@ web_table <- function(data
                       , buttons = NULL
                       , file_name = "file"
                       , scrolly = NULL
+                      , columnwidth = "200px"
+                      , width = "100%"
                       ){
   
 # -------------------------------------------------------------------------
@@ -43,7 +47,7 @@ web_table <- function(data
   
   where <- NULL
   
-  if(is.null(scrolly)) scrolly <- "60vh"
+  if(is.null(scrolly)) scrolly <- "45vh"
   
   if (is.null(buttons)) {
     
@@ -58,9 +62,10 @@ web_table <- function(data
   
 # -------------------------------------------------------------------------
 
-  data %>% 
+  table <- data %>% 
     mutate(across(where(is.numeric), ~round(., digits = digits))) %>%
     datatable(extensions = ext
+              , width = width
               , rownames = rnames
               , options = list(
                 dom = 'Bt' # "Bti"
@@ -69,15 +74,18 @@ web_table <- function(data
                 , scroller = TRUE
                 , scrollX = TRUE
                 , scrollY = scrolly
-                
-                , columnDefs = list(list(width = '200px'
+                , autoWidth = TRUE
+                , columnDefs = list(list(width = columnwidth
                                          , targets = "_all"))
                 
                 , initComplete = DT::JS(
                   "function(settings, json) {",
+                  "$(this).addClass('compact');",
                   "$(this.api().table().header()).css({'background-color': '#000', 'color': '#fff'});",
                   "}")
                 )
               , caption = caption)
+  
+  table
 
 }